Change add to shopping cart to use Java
