+++ src/main/java/ShopifyProj/Controller/FirebaseController.java
@@ -1,13 +1,16 @@
 package ShopifyProj.Controller;
 
+import ShopifyProj.Model.TempShop;
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.FirebaseOptions;
 import com.google.firebase.database.*;
+import org.springframework.ui.Model;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.util.concurrent.CountDownLatch;
 
 public class FirebaseController {
     private static FirebaseController inst = null;
@@ -43,5 +46,32 @@ public class FirebaseController {
         }
 
         return (dbInst);
+    }
+
+    public static TempShop getShopFromID(String aShopID) {
+        final TempShop[] shop = {null};
+        CountDownLatch wait = new CountDownLatch(1);
+        FirebaseController.getInstance().getReference("store/" + aShopID).addListenerForSingleValueEvent(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                if (dataSnapshot.exists()) {
+                    shop[0] = dataSnapshot.getValue(TempShop.class);
+                }
+                wait.countDown();
+            }
+
+            @Override
+            public void onCancelled(DatabaseError databaseError) {
+                System.out.println("The read failed: " + databaseError.getCode());
+                wait.countDown();
+            }
+
+        });
+        try {
+            wait.await();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return shop[0];
     }
 }
